% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{euc_dist}
\alias{euc_dist}
\title{Fast euclidean distance}
\usage{
euc_dist(m)
}
\arguments{
\item{m}{matrix}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function returns Euclidean distance between all combination of rows of a data matrix. Much faster than 'dist' when you use large matrix (n> 100.000.000).
}
\examples{
M <-  matrix(rnorm(100), nrow = 5)
euc_dist(M)
}
