% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_stats.R
\name{fastzonal}
\alias{fastzonal}
\title{Fast implementation of zonal statistics}
\usage{
fastzonal(x, z, stats, digits = 0, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A number.}

\item{z}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Fast implementation of zonal statistics
}
\examples{
add(1, 1)
add(10, 1)
}
