% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_stats.R
\name{my_zonal}
\alias{my_zonal}
\title{Add together two numbers.}
\usage{
my_zonal(x, z, stat, digits = 0, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A number.}

\item{z}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
add(10, 1)
}
