% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_stats.R
\name{zonal_pipe}
\alias{zonal_pipe}
\title{Add together two numbers.}
\usage{
zonal_pipe(r, z, stats, filename = NULL)
}
\arguments{
\item{r}{A raster file for which one wishes to compute zonal statistics}

\item{z}{A vector layer (e.g. a shapefile) containing one polygon for each zone used to derive zonal statics (e.g. admnistrative areas)}

\item{stats}{The function applied on all pixels of an area (i.e. the zonal statistics)}
}
\value{
the path where to write the vector layer with the result of the zonal statistics. If set to NULL returns the vector layer object (already).
}
\description{
Add together two numbers.
}
\examples{
output <- zonal_pipe(belgium, precipitation, stats = "mean")


}
