% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_stats.R
\name{zonal_pipe}
\alias{zonal_pipe}
\title{Add together two numbers.}
\usage{
zonal_pipe(zone_in, raster_in, shp_out = NULL, stat)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
zonal_pipe(1, 1)
zonal_pipe(10, 1)

}
